chore: move decodePermisson and encodePermissions to ERC725 class
