+++ docs/classes/ERC725.md
@@ -398,7 +398,7 @@ await myERC725.getOwner('0x3000783905Cc7170cCCe49a4112Deda952DDBe24');
 
 ▸ `encodePermissions(permissions): string`
 
-Encode permissions into a hexidecimal string as defined by the LSP6 KeyManager Standard.
+Encode permissions into a hexadecimal string as defined by the [LSP6 KeyManager Standard](https://docs.lukso.tech/standards/universal-profile/lsp6-key-manager).
 
 ### Parameters
 
@@ -410,12 +410,12 @@ Encode permissions into a hexidecimal string as defined by the LSP6 KeyManager S
 
 `string`
 
-The permissions encoded as a hexidecimal string as defined by the LSP6 Standard.
+The permissions encoded as a hexadecimal string as defined by the [LSP6 Standard](https://docs.lukso.tech/standards/universal-profile/lsp6-key-manager).
 
 ### Example
 
 ```javascript
-encodePermissions({
+ERC725.encodePermissions({
   CHANGEOWNER: false,
   CHANGEPERMISSIONS: false,
   ADDPERMISSIONS: false,
@@ -433,7 +433,7 @@ encodePermissions({
 
 // Any Ommited Permissions will default to false
 
-encodePermissions({
+ERC725.encodePermissions({
   CHANGEPERMISSIONS: true,
   SETDATA: true,
 }),
@@ -444,11 +444,11 @@ encodePermissions({
 
 ---
 
-## decodePermissions
+## decodePermission
 
-▸ `decodePermissions(permission): Object`
+▸ `decodePermission(permission): Object`
 
-Decodes permissions from hexidecimal as defined by the LSP6 KeyManager Standard.
+Decodes permissions from hexadecimal as defined by the [LSP6 KeyManager Standard](https://docs.lukso.tech/standards/universal-profile/lsp6-key-manager).
 
 ### Parameters
 
@@ -460,12 +460,12 @@ Decodes permissions from hexidecimal as defined by the LSP6 KeyManager Standard.
 
 `Object`
 
-Object specifying whether default LSP6 permissions are included in provided hexidemical string.
+Object specifying whether default LSP6 permissions are included in provided hexademical string.
 
 ### Example
 
 ```javascript
-decodePermission('0x0000000000000000000000000000000000000000000000000000000000000110'),
+ERC725.decodePermission('0x0000000000000000000000000000000000000000000000000000000000000110'),
 /** 
 {
   CHANGEOWNER: false,
@@ -481,7 +481,7 @@ decodePermission('0x000000000000000000000000000000000000000000000000000000000000
 }
 */
 
-encodePermissions('0x000000000000000000000000000000000000000000000000000000000000000a'),
+ERC725.decodePermission('0x000000000000000000000000000000000000000000000000000000000000000a'),
 /** 
 {
   CHANGEOWNER: false,
